new_ai_backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
ollama>=0.1.5
numpy>=1.26.0
scikit-learn>=1.3.1 
playwright>=1.40.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
google-search-results>=2.4.2
requests>=2.31.0
asyncio>=3.4.3
aiohttp>=3.9.1
google-api-python-client>=2.107.0
